Remove the scoping of region via the ENV variable
This appears to be at the heart of the issues you were having I think.
This was intended to be a short term optimisation, but if it's causing
too much uncertainty we should pull this out.

We can come back to performance and questions over whether scoping or
caching are good ideas/required for a nice user experience.

This removes that problem but retains the puppet resource functionality.