Fix for getting security group in acceptance tests
The previous behaviour would grab the first security group, in the case
of the test named default. That happened to be the non-vpc group. The
addition of VPC support to instance means that it now grabs both the
default groups, but as a non-vpc instance this won't work. This change:

* Adds an acceptance-non-vpc group just for testing
* Bumps the timeout for the initial test as I saw a few failures (maybe because I'm further away)