Initial support for VPC resources, including:
* VPC
* Subnet
* Internet gateway
* Route table

Also includes modifications to instances and security groups to work
with VPC.

I've been using this scenario as a basis to get to the first working
version of this support:
http://docs.aws.amazon.com/AmazonVPC/latest/UserGuide/VPC_Scenario1.html

At this point it would be useful to review I think from a functionality
point of view. It's a bigger than ideal diff and it would be better to
get it in sooner rather than later.