The destroy example didn't run in one go
This demonstrates using wait_until to allow for working around the
dependencies that require multiple runs to delete.

This will fail the spec tests at the moment due to the extra API call
for delete. If we like this approach then I can fix up the test, if you
want to merge it sooner then feel free to mark the delete test as
pending.

I think doing this, while it slows deletes down, is probably worth it. I
don't think we should wait everywhere, the parallelisation of everything
coming up at once is too useful in my view. Worst case we could make
this toggleable with a shared param.