Use a flag to run tests under beaker
Rather than have two parallel test suites, this proposal allows toggling
tests to be run using beaker with an ENV variable.

This is currently a working demonstration with a number of hard coded
bits, in particular:

* Uses Puppet rather than allowing for PE as well
* Uses a single node as the master and agent
* Uses Vagrant as the backend, rather than allowing for Vagrant and vmspooler

These are all easily changable, but want to get feedback on the
approac first.