Improve formatting of tags when output in logs
Previously tags would be output in the order they were entered or
returned from the AWS API, which might be different. The string
representations could also use single quotes or double quotes. All of
which made parsing changes by eye difficult. This change normalises both
sides of the is and should to be the same, as well as ordering by key
name.