Allow creating an instance in the stopped state
This creates the instance if it doesn't exist, and then puts in the
instance in the stopped state.

Note that an instance has to be running before it can be stopped, so we
now wait for the instance to be running in create. I've done this in
create rather than in stop because of the occasional failures to tag
instances. I'd rather wait until the instance is in pending, but the API
doesn't currently support that:
https://github.com/aws/aws-sdk-core-ruby/issues/199