Make TXT record interface more Puppet like
TXT records are always wrapped in quotes, which leads to passing values
like '"value"' which makes sense in DNS land but less sense in Puppet.
This change improves that interface, adding the additional inner quotes
if they aren't already present.

Also allow values to be double quoted. This allows for the explicit DNS case,
which matches how the AWS console and other DNS tools would work.