Work around failures to tag new instances
Due to either network failures (over the internet) and/or eventual
consistency issues in AWS, it's possible to start an instance and then
fail to tag it. The API doesn't expose a tranactional endpoint for this.

Unfortunately, names of instances are just tags. So in the case tagging
fails we can't retrieve information via name later.

We have tried simply waiting for the instance to be running (which
minimises the eventual consistency issues) but have still seen the odd
failure. This patch replaces that wait (a good thing, instance creation
is now quicker from Puppet's point of view) with retries of the tagging
calls in case of failure.