Retry gateway deletion while waiting for detachment
The SDK doesn't have a native waiter for this state, so we retry several
times with a back-off for the specific case where the gateway hasn't
detached (specified by the exception caught). This does still often take
up to 60s, but at least in the case it doesn't this is faster.