Avoid acceptance test failures due to reuse of IP address
The previous hard coding of the IP address for the customer gateway
meant that multiple runs of the acceptance tests in parallel could cause
a failure of one. Specifically this relates to
https://aws.amazon.com/premiumsupport/knowledge-center/conflict-among-chosen-gateway-ip-addresses/

This commit uses a random ip from a well populated IP range, which
should reduce the occurance of this issue.