(CLOUD-232) Add VPC security group support
The security groups in VPCs must be managed by their IDs, so I added
`id` to the ec2_securitygroup type and updated the provider to only deal
with the IDs.

The `vpc_name` property for ec2_securitygroup was added to create a SG
in a VPC. This may only be done on SG creation, so I created a setter
method that raises an error if it is not insync after create.

The ec2_vpc type now returns its ID. The provider was already setting
it in `instances`.