Allow specifying default as the security group
Default is the standard name for auto-generated security groups in VPCs,
which means we need to filter on the VPC in which the ELB is in (via the
specified subnet) to determine the correct security group in this case.

We also need to deal with the case where we don't specify anything and
so end up with default, in this case we don't want to trigger an
out-of-sync.