Make clearer the rules around passing multiple domain names for options
Worth norting from the documentation:

Some Linux operating systems accept multiple domain names separated by
spaces. However, Windows and other Linux operating systems treat the
value as a single domain, which results in unexpected behavior. If your
DHCP options set is associated with a VPC that has instances with
multiple operating systems, specify only one domain name.

This change ensures that if the domain name is actually a space
separated list then we convert it to a set in order to determine whether
it's what we asked for.