Avoid throwing exceptions in prefetch
I have a guess this accounts for at least some orphaned resources left
after test runs. The story goes like this:

* Get a list of a resources (say instances)
* For each of those instances, we trigger further requests to get
  the names from the ids (say for related security groups)
* This was previously done using the *_ids param to describe_*
* This formulation will throw an exception if a resource with that id
  exist
* If prefetch fails, the Puppet run continues (this is fixed in 4.0)
* But the entire prefetch for that resource fails
* Resources won't correctly deal with exists?
* We could create extra resources or not delete others

This happens infrequently in the small, because it generally relies on
something else to be operating on the resources during the operation -
ie. between the initial describe_* call, another call deletes an
unrelated resource. But in the large that happens enough, especially
with the test suite.

This change swaps out the *_ids param for a filter on the *-id property.
This does exactly the same thing, but doesn't throw an exception. So now
we can simply deal with an empty array as needed.