Memoize calls made to AWS by ID
This commit moves a bunch of functionality from the providers to the
base class. This centralisation allows for some basic caching.
Specically we regularly look up the name (often in a tag) of a resource
by it's AWS id.

This cuts down on the number of API calls made, especially as the
resource count increases.