Prevent issues accessing unattached gateways
VPN gateways in particular can take time to attach, which results in a
potential race condition - often with dependent route tables failing due
to not being able to find the gateway in the correct state.

This commit addresses that by waiting for a period of time for the
gateway to attach before marking the resource present. To be safe this
waits for attached rather than attaching.