Deal with blank name tags for all resources
This was previously fixed for instances only but this commit generalises
the fix to other resources. Tags do not have to contain content, so it's
possible for an account to contain resources with a Name tag which is
blank. Previsould we checked for the presence of name, and not whether
it was empty.