Don't make queries that we'll ignore the results of anyway
We only manage resources with names, in instances we already reject
anything that doesn't have one. So we don't need to make additional
requests to find out details of related resources for things we don't
care about.

This commit causes the to_hash methods to return early if a name isn't
found.