Filter on vpc-id for sec groups in EC2-VPC
When adding security_group ingress rules to a security group resource
upon creation, a VPC-only account may have multiple groups with the
same name, often 'default', to select from. Without this change, the
ec2_securitygroup provider will simply select the first that it finds.
If the one it selects is not in the VPC to which the security group
puppet resource was added, it will raise an error and fail to add the
ingress rule. This change ensures that the correct security group for
the resource's ingress rule is selected by retrieving the VPC ID of the
security group that was added and using it as part of the filter for
selecting possible security groups to use for the ingress rule.

Paired-with: <ethan@puppetlabs.com>