Remove the additional query for VPC lookup for security groups
Security groups can only associate with those in the same VPC. See
the following gist for an example of what happens when you try:

https://gist.github.com/garethr/81ef20bba61e89c037e6

That means it's only valid for the security group we're connecting with
to have the same VPC id as the security group we're changing. So by
stashing the vpc_id locally we can avoid the extra query to look it up.

This also accomodates VPCs in Classic accounts.