(CLOUD-179) Terminate run on inconsistent AWS data
If we query AWS for a resource (by id) that doesn't exist we will raise
an exception. Due to PUP-3656 that prevents prefetch but lets the
provider continue with incorrect data.

This changes that behaviour to stop the provider run if we discover
inconsistent data. It also provides information to help debug the
problem - although in many cases running a second time will fix the
issue.

Common causes of inconsistency are:

* eventual consistency / propagation of resource data within AWS
* other clients acting on the API at the same time

This change also allows rate-limiting errors to immediately cause the
run to stop.

Paired-with: <colleen@puppetlabs.com>
Paired-with: <zach.reichert@puppetlabs.com>
Paired-with: <ethan@puppetlabs.com>