Drastically reduce number of calls for subnets in large installs
The number of subnets is generally less or much less that the number of
instances. Previously we requested subnet information for each instance
in prefetch, this optimisation gets data about all subnets and uses it
as needed. From basic testing on a medium sized install this reduced the
prefetch time by as much as a third. On a large install this improvement
is much greater.