(MAINT) We shouldn't care about the return order of availability zones
This test just started failing in Jenkins. The test is overly
prescriptive in that we want both availability zones returned but we
don't care about the order. On the other hand this has to my knowledge
never failed before so it's fixed order, just whatever is fixing it has
changed.

This change makes the test more robust, simply allowing both possible
options.