Limit elastic IP association to pending, running instances
EC2 takes some time to clean up terminated instances. When replacing an instance, elastic association on the new instance often fails because the old instance is still hanging around in a "terminated" state.

This change limits instance lookup to pending or running instances for the purpose of elastic IP association.