(CLOUD-412, CLOUD-411) Support specifying backup retention
It was pointed out that the AWS console defaults the backup retention to
30 days, however the API defaults it to 1 day. Because we we're exposing
that parameter this was both impossible to change and a bit invisible.

This change adds the ability to manage the backup retention period for
new RDS instances.

In the process of testing that change to ensure it was idempotent I
added the relevant (missing) test. This passed for the new property but
failed for another one (allocated_storage). This was reported separately
in CLOUD-411. This commit fixes that too.