[BUGFIX] Error: Could not set 'present' on ensure: unexpected value at params[:subnet_group_name]
Trying to create an `rds_instance` with the `db_subnet` parameter failed with
unexpected value at params[:subnet_group_name].

The AWS CLI manual for `rds create-db-instance` specifies the parameter as `[--db-subnet-group-name <value>]`.

Updating the parameter to the correct name fixes the error and results in
correct creation of an RDS instance within the VPC subnet group.