Add support for region and proxy facts
When using this module during a Puppet agent run (such as on a dedicated AWS arbiter node), the environment variables AWS_REGION and PUPPET_AWS_PROXY are not available to Puppet.  This adds support for facts named aws_region and puppet_aws_proxy that can be used for the same purpose.