(MAINT) Fix issue where an exception triggers another exception
This is a copy-paste bug. Route53 is a global service (it's DNS), unlike
all the other services which are scoped to region. This is rarely seen,
but a user managing a large number of records hit it a few times. This
should mean when the exception occurs a useful error message is produced
rather than getting Error: Failed to apply catalog: undefined local
variable or method `region' for.