(FM-4849) Properly manage subnets and availability zones on loadbalancers
Subnets and availability zones are interdependent entities: changing the subnets
changes the availability_zones and vice versa. The elb_loadbalancer type
only allows one of the two properties to be set. There are
three modes of operation here:

1. EC2-Classic: there are no subnets, and the LB is configured using availability_zones
2. VPC, using defaults: the LB is configured using availability_zones, using the default subnets for each zone
3. full VPC: the loadbalancer is configured using subnets; availability_zones are ignored

Before this EC2-Classic and VPC using defaults modes were not working,
because the API would return both availability zones and subnets, confusing
the default implementation. With the changes in this commit, availability_zones
are only used during create, as documented in the README, and implemented before,
but subnets are properly ignored. In the case of subnets being used to configure
the loadbalancer, availability_zones are properly ignored.