provider/cloudwatch_alarm: do not convert arns into autoscaling groups
The way I detect that it's a arn is by checking whether the string
starts with arn:aws. There are probably better ways of doing that. We
could be more specific and support only certain kinds of arns, for
example.

I failed to change/prepare a test for this change. I'd love to include a
test in this change, if someone could point me to the right direction.

This is related to #227.