Add VPC support for RDS instance creation
Without this change, it is not possible to create an RDS instance
attached to a VPC security group.  This work extends the current
rds_instance provider to include the handling of the vpc_security_groups
parameter and adds the necessary configuration for the API calls to
manage the state as requested.