Initial Ec2 Volume management
Without this work, managing volumes independent of Ec2 instances is not
possible using this module.  This is the preliminary work to get
Ec2 Volume management operational under Puppet.

Expect that this will be followed up with adjustments to the
ec2_instance type to allow for attachment of volumes that are managed
outside of the ec2_instance type.  Future work will allow volumes to be
carried from one instance to another as instances are decommissioned,
etc.