Add alias_target property for Route53 types
The alias_target of a route53 resource record is a stand in value that
allows resources to shift behind the scenes without requiring action on
the part of the user.  Without this change, we are unable to set or
retrieve the alias_target of a route53 record.  Here we simply add a
alias_target property to enumerate the value if its found to exist.