Replace "checking" info with debug
Checking to see if a resource already exists in AWS is something that is
common enough that printing an info message each run for each resource
is not informative.  Knowing what the agent is doing during a catalog
application would be useful for debugging.  As such, here we convert the
Puppet.info to Puppet.debug for the providers' exists?() method to
suppresses the message.