Refactor ECS task container merge code
This work adjusts the code that compares the container_definition
parameter on the ecs_task_definition type to allow the compared
containers to be passed in, rather than looked up from the
@property_hash and @flush_hash variables.  This allows us to make use of
this comparison in the in_sync?() method of the type to determine if
changes need to be made.

The advantage to this approach is to allow missing options within a
container definition on the resource not to trigger an in_sync() failure
and thus flush() call, while the options in question do indeed exist in
ECS.  This will allow other tools (CI, etc) to modify options on a
container like 'image', while puppet can handle the rest without
conflict.