Implement retry for ELB request limits
During the enumeration of ELB resources, several calls to the AWS API
are made in a short amount of time.  In some cases, this can cause the
API query limits imposed by AWS to be reached.  This is a problem
because puppet fails the resource and all dependent resources when this
occurs.  This change rescues the exception thrown by the aws-sdk
specific to this condition and sleeps a period of time before retrying.