Fixes for iam_group absent
Without this change, Puppet does not successfully remove a group if
there are members of the group, or policies are attached.  This work
updates the destroy method on the iam_group provider to perform the
necessary lookups for what is attached to the group and perform the
membership eviction and policy detachment necessary to delete the group.