Add param to skip image comparison on ECS container
Without this change, the behavior of puppet is to take all container
properties into consideration during comparison what is with what should
be.  Normally this makes a great deal of sense, but becomes a problem in
environments where an external tool (as part of CI) modifies the image
of a container to reflect the current desired build.  This results in
puppet overwriting the image due to correctly taking the responsibility
of managing the entire container definition.

This change adds a new parameter to the ecs_task_definition type called
'replace_image', which when set to false, will copy the existing image
value from the existing container into the container that puppet reports
"should" be.  This has the effect of ignoring the image that has been
set in Puppet, while still allowing the comparison for all other aspects
of a container to retain existing comparison functionality.