Add availability_zone to RDS provider
Without this change, the rds_instance provider doesn't enumerate or
create the RDS instance in the requested availability zone.  This is a
problem because Puppet takes the property and does nothing with it,
potentially causing a dangerous deployment.  Here we append
availability_zone to the list of properties for creation and enumeration
of the RDS instances.