Add support for modifying key attributes of RDS
Without this change, the RDS instances that are created with the
rds_instance type are not modifyable.  Puppet detects the changes
correctly, but does nothing to update the real world state.  Here we
take a few important parameters that one might want to modify on the
running rds_instance resources, and flush the changes from the setter
methods.