Improve ELB enumeration performance and stuff
Without this change, when we need to generate a Hash of all the
information about a particular ELB during the resolution of instances,
and in the process, convert any security group, subnet, and instances
that the ELB is attached to from ID strings into resource names,
subsequent calls are made to the API for each of these enumerations.
This presents a problem due to the number of calls made to the API,
which is three calls to the Ec2 API per ELB.  This is inefficient,
because we have an autorequire that ensures the instances calls on the
relevant puppet resources are already enumerated, meaning that much of
the data we require in the ELB hash generation is already present in
memory, but we make the API calls anyway.  This is only true for
resources under Puppet management, as resources not under puppet
management are thrown away during the instances call for a given
provider.  In environments with lots of ELBs, this can mean that the API
limit on Ec2 is reached, and puppet fails to run, which prevents scaling
and clean runs of Puppet, thus raising the occurrences of swearing.

This work changes the behaviour of the ELB enumeration to first look in
the catalog catalog for the resources that might match, and see if the
relevant data is present before we make the calls to the API.  The
result is that for resources under Puppet management, no extra calls
need to be made to the Ec2 API, thus improving performance and reducing
the risk of reaching the Ec2 API limits.  This also means that there is
an incentive for moving resources under Puppet management, as even those
ELBs that are not managed by Puppet, but are present in the environment
will drag performance down by requiring subsequent calls for the
security groups, subnets, and Ec2 instances related to the particular
ELB.