Add support for security_group changes on ec2_instance
Without this work, the security_groups parameter on the ec2_instance
type is marked as read-only.  Here we remove the security_groups
parameter from the read-only list and implement a setter method and the
necessary flush logic to modify the security groups attached to an
interface.

This is currently only implemented for instances with a single
interface, as the modification for what security groups are associated
with an instance is a change needed on the interface, and not the
instance.  This is a bit confusing because the interface security groups
are represented on both the main instance object, and the interface
object.  For this reason, a new read-only parameter is also added here
to collect information about the interfaces interfaces to ensure that we
are only modifying the instances we know how, as well as collecting the
data to get the interface ID for modification in the flush method.