Begin filtering sensitive data during VCR recording
Without this change, module testers are left to their own methods to
ensure that they are not leaking any sensitive material when
contributing to this module.  This is not good, as a) we want more tests
and b) we would like not to expose data from contributor environments
when submitting changes.

Here we take a stab addressing this concern by configuring VCR to filter
out the AWS_ACCESS_KEY_ID.  This alone is not enough to be complete, but
it begins the road of making testing simpler on the part of those who
wish to contribute to the module.  Now we can amend the documentation to
include information about exporting the necessary environment variables,
not providers specs don't have to modified last minute to avoid
committing credentials.