Add refresh for Ecs_service resources
Without this change, the ECS service is unable to track changes to the
task.  This leaves the code in a state where the task definition is
updated, but the service is not refreshed to be running on that code.
Here we implement the ability for the Ecs_service resources to subscribe
to Ecs_task_definition resources.