Allow security groups mutual peering
Without this change, two security groups cannot peer with each other. The first
security group will be created, but will fail to add rules since the second
security group does not exist yet. Further security groups will not be created
due to the first security group failing and being a dependency. This skips
rules that refer to security groups that do not exist.