Enable autorequire for iam_role for policy attachment
Without this change, the autorequire for iam_role is not enabled for the
iam_policy_attachment.  This was originally commented because the
iam_role code had not been merged.  Now that the code has been merged,
we can enable this here to allow the automatic and correct ordering for
IAM resources.