Begin KMS management
Without this change, there is no functionality in this module to manage
the lifecycle and policy of a KMS key in amazon.  Here we add a simple
type and provider with just enough to create, destroy, and manage the
policy.  There are testing updates for the normalization code here,
since the KMS policy exercises data structures that are not exercised
elsewhere, so we extend the use case a bit of the normalize method
calls.