Improve ELB listener detection
Without this change, an ELB with no listeners causes a stacktrace due to
an incorrect detection of existing listner.  Here we adjust the logic to
account for an empty array, by checking that we have at least 1 listener
before we proceed inspecting the delta between the desired listeners and
the actual listeners.