Fix new RDS instance tagging
Without these changes, RDS instance creation breaks on tagging the new
instance. This saves the response from RDS instance creation so the generated
ARN can be used to tag the RDS instance. It removes use of `tags_for_resource`
since that expects the tags property to be named `tags` instead of `rds_tags`.