Improve creating RDS instances from snapshot
Without this change, if the resource describing an RDS instance includes a
snapshot identifier and any of several properties not used in building an RDS
instance from a snapshot, the creation from snapshot would fail. This prunes
these values to allow creation.

This also changes RDS instance creation to include tagging in the same call.