Improve name resolution method isolation and debug output
This change moves some of the logic for retrieving object names into their own
method to help readability.  Also added here is improved debug output about why
a call to AWS is being made for each of the objects that fail to resolve from
the catalog.  This should help users identify which objects need to be pulled
into the catalog to improve performance.